requirements.txt
streamlit==1.35.0
opencv-python==4.9.0.80
pytesseract==0.3.10
Pillow==10.3.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
rapidfuzz==3.9.3
